Split CommunicateBoundingBox file and class
